requirements.txt
sqlalchemy
sqlalchemy-utils
psycopg2-binary
sqlacodegen-v2
python-dotenv
python-dateutil
flask